chapter_7/recipe_4/src/requirements.txt
requests==2.26.0
google-cloud-secret-manager==2.7.2
